finished 1.44
Kept trying to pass in an argument to smooth thinking we need to execute
it immediately but that just result in the wrong answer.
Smoothed takes a function and returns a lambda.
Was much easier to figure out once i defined n-fold-smoothed instead of
trying to compute directly.
Need to listen to Stepanov and think harder